Remove unneeded `test_min_allocatable_resources`
This test was introduced to test the correct behavior of a custom
setting for Mesos' `min_allocatable_resources` flag. We have since
removed this custom flag setting which both caused this test to fail
while also removing any need for this test, so we are removing the test
in this patch.