Added linting for DC/OS commits.
This commits adds a configuration file for the `pre-commit`
git hook lifecycle management framework. It is used to
run various linting checks on all git commits.

In particular, initially we check that all commit messages
start with an uppercase letter, no line exceeds 72 characters,
that all executables have valid shebangs, and that no non-ascii
characters are introduced accidentallu.

To enable pre-commit for this repository, first pre-commit must
be installed

    pip3 install pre-commit

and afterwards the following commands must be run once in the
the root folder of the git repository:

    pre-commit install --install-hooks --hook-type pre-commit
    pre-commit install --install-hooks --hook-type commit-msg

Subsequent changes to the linter configuration will be picked
up automatically.

For more information, visit pre-commit.com