Remove CockroachDB poststart commands
If an ExecStartPost command times-out, then the unit is killed,
even when the command starts with `-`.  As the post-start steps
here are also executed by a periodic timer, we remove them here
to prevent the post-start timing out while waiting for Cockroach
to start, and then preventing Cockroach from starting.