Run Windows tests withouth Linux private nodes
In order to really test Windows agents we need to ensure
that they are hit by tests we are running. To do that we need to remove
private linux nodes and run tests only on Windows node.

We can't remove linux public nodes since this results in tests not
starting at all.

This PR fixes dcos-diagnostics tests to accept clusters without Linux
private nodes and to use Docker App that's running on Windows and Linux.

This PR skip couple of metrics tests since statsd metrics are not
working on Windows and apps useds in tests do not run on Windows.

Refs: D2IQ-66051
Fises: D2IQ-65996