Do not require diagnostics on destroy.
This leads to weird situations when we want to destroy cluster
but first we need to create a diagnostics bundle for it.
It's especially anoing when you want to stop cluster creation and
free resources but first you need to finish a cluster and then destroy
it.

After merging this change we need to add step to our CI to generate
diagnostics bundle even if previous step failed.