Stop all services at once
Currently we stop servies one by one. This may lead to a situation when
we stop service before we stop it's socket activator. So if any packet
gets sent to that socket, the service will be activated.
To prevent this we need to stop all services in a single command.

Fixes: https://jira.d2iq.com/browse/COPS-6512